pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "openai>=1.97.1",
    "psycopg2-binary>=2.9.10",
    "flask-wtf>=1.2.2",
    "sqlalchemy>=2.0.41",
    "werkzeug>=3.1.3",
    "pymupdf>=1.26.3",
    "pillow>=12.1.1",
    "reportlab>=4.4.3",
    "requests>=2.32.4",
    "flask-security-too>=5.6.2",
    "flask-mail>=0.10.0",
    "python-dotenv>=1.1.1",
    "fitz>=0.0.1.dev2",
    "argon2-cffi>=25.1.0",
    "flask-security>=5.6.2",
    "flask-dropzone>=2.0.0",
    "boto3>=1.39.12",
    "oauthlib>=3.3.1",
    "scipy>=1.16.0",
    "numpy>=2.3.1",
    "psd-tools>=1.10.8",
    "pyjwt>=2.10.1",
    "flask-dance>=7.1.0",
    "urllib3>=2.6.3",
]

[project.optional-dependencies]
dev = [
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
    "mypy>=1.17.0",
    "flake8>=7.3.0",
    "black>=25.1.0",
    "pytest>=8.4.1",
    "isort>=6.0.1",
]
